+++ libs/domains/src/user/application/queries/find-login-user-by-social-account/find-login-user-by-social-account.query.ts
@@ -1,6 +1,6 @@
 import { IQuery } from '@nestjs/cqrs';
 
-export class UserGetBySocialAccountQuery implements IQuery {
+export class FindLoginUserBySocialAccountQuery implements IQuery {
   constructor(
     public readonly provider: string,
     public readonly socialId: string,
